pipeline/requirements.txt
faust-streaming
google-cloud
google-cloud-bigquery
web3
confluent-kafka
